@@ -1,17 +1,44 @@
+const CACHE_NAME = 'v1';
+
 self.addEventListener('install', function (event) {
     event.waitUntil(
-        caches.open('v1').then(function (cache) {
+        caches.open(CACHE_NAME).then(function (cache) {
             return cache.addAll([
+                '/',
                 '/index.html',
                 '/static/js/bundle.js',
-                // List other assets
+                // Add paths for other static assets
             ]);
         })
     );
 });
 
+self.addEventListener('activate', function (event) {
+    event.waitUntil(
+        caches.keys().then(function (cacheNames) {
+            return Promise.all(
+                cacheNames.map(function (cacheName) {
+                    if (cacheName !== CACHE_NAME) {
+                        return caches.delete(cacheName);
+                    }
+                })
+            );
+        })
+    );
+});
+
 self.addEventListener('fetch', function (event) {
-    event.respondWith(caches.match(event.request).then(function (response) {
-        return response || fetch(event.request);
-    }));
+    event.respondWith(
+        caches.open(CACHE_NAME).then(function (cache) {
+            return cache.match(event.request).then(function (response) {
+                if (response) {
+                    return response;
+                }
+                return fetch(event.request).then(function (response) {
+                    cache.put(event.request, response.clone());
+                    return response;
+                });
+            });
+        })
+    );
 });
